Require Chef 12.14+ and remove apt/yum dependencies
Chef 12 is end of life so there's really no need to continue to pull in
the apt and yum dependencies here. Instead of maintaining the 12.1
requirement bump it to 12.14 which came out Sept 2016. Even if you
haven't upgraded to Chef 13 there's no reason not to be on one of the
later Chef 12 releases.

Signed-off-by: Tim Smith <tsmith@chef.io>